@@ -8,8 +8,10 @@ import { ItemsListPageComponent } from './items-list-page/items-list-page.compon
 import { LandingPageComponent } from './landing-page/landing-page.component';
 import { PartyPageComponent } from './party-page/party-page.component';
 import { PlacePageComponent } from './place-page/place-page.component';
+import { SharePageComponent } from './share-page/share-page.component';
 import { ShoppingItemComponent } from './shopping-item/shopping-item.component';
 import { ShoppingListPageComponent } from './shopping-list-page/shopping-list-page.component';
+import { CalculatePageComponent } from './calculate-page/calculate-page.component';
 
 const routes: Routes = [
   { path: '', component: LandingPageComponent },
@@ -19,11 +21,13 @@ const routes: Routes = [
   { path: 'items/:id', component: ItemsListPageComponent },
   { path: 'place', component: PlacePageComponent },
   { path: 'shoppingList', component: ShoppingListPageComponent},
-  { path: 'categoryList', component: CategoryListPageComponent}
+  { path: 'categoryList', component: CategoryListPageComponent},
+  { path: 'share', component: SharePageComponent},
+  { path: 'calculate', component: CalculatePageComponent}
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forRoot(routes, { relativeLinkResolution: 'legacy' })],
   exports: [RouterModule]
 })
 export class AppRoutingModule { }
