@@ -11,7 +11,6 @@ import { Category,CategoryBody } from '../party'
 export class CategoryListPageComponent implements OnInit {
 
   partyId: String = localStorage.getItem('partyId');
-  addNewVisible: boolean = false;
 
   categoryList: Category[];
   
@@ -27,19 +26,20 @@ export class CategoryListPageComponent implements OnInit {
   }
 
   createCategory(categoryName: String){
+    
+    
     let category= new Category;
     category.name=categoryName;
+    category.items=[];
     this._ppRest.putCategory(this.partyId,category).subscribe(
-      data => {
-        this.categoryList.push(category);//ni glih dobr zato ker nima ID-ja in ce hocs brisat al karkol je pizdarja
-      },
-      error => {
-        console.log('error');
-      }); 
-  }
-
-  addCategory(){
-    this.addNewVisible = !this.addNewVisible;
+      (data:Category) => {
+       if(data != null || data != undefined){
+         this.categoryList.push(data);
+       }
+     },
+     error => {
+       console.log(error);
+     });
   }
 
 
