@@ -26,34 +26,34 @@ export class ItemsListPageComponent implements OnInit {
         
     this._ppRest.getCategoryById(this.partyId,this.categoryId).subscribe((result:Category)=>{
       this.category=result;
+      this.itemList=result.items;
     });
   }
 
   ngOnInit(): void {
   }
  
-  createItem(name:String,description:String,price:number){
+  createItem(name:String,description:String,price:string){
     
     let item = new Item;   
-    //item.itemId = this.itemList.length;
     item.name=name;
     item.description=description;
-    item.price=price;
+    item.price=Number.parseFloat(price);
     item.quantity=this.addNewQuantity;
+    this.addNewQuantity=0;
     this.itemList.push(item);
 
     this.category.items=this.itemList;
-    
+ 
     this._ppRest.postCategory(this.partyId,this.category).subscribe(
       data => {
-        //this.reloadCurrentRoute()
-        console.log("dela baje");
+        console.log(data);
       },
       error => {
-        console.log('error');
+        console.log(error);
     });
-
     
+
     
   }
 
@@ -64,18 +64,38 @@ export class ItemsListPageComponent implements OnInit {
     if(this.addNewQuantity > 0) this.addNewQuantity--;
   }
 
-  on(){
-    document.getElementById('overlay').style.display="block"
-  }
-  off(){
-    document.getElementById('overlay').style.display="none"
-  }
+  // on(){
+  //   document.getElementById('overlay').style.display="block"
+  // }
+  // off(){
+  //   document.getElementById('overlay').style.display="none"
+  // }
 
-  shopping(){
-    this.router.navigate(['/shoppingList']) 
-  }
+  // shopping(){
+  //   this.router.navigate(['/shoppingList']) 
+  // }
+  
   party(){
     this.router.navigate(['/categoryList']);
   }
+  
+  deleteItem(item:Item){
+    this.itemList=this.delete(this.itemList,item);
+    this.category.items=this.itemList;
+    this._ppRest.postCategory(this.partyId,this.category).subscribe(
+      data => {
+        console.log(data);
+      },
+      error => {
+        console.log(error);
+    });
+  }
 
+  delete(array,item){
+    const index = array.indexOf(item, 0);
+    if (index > -1) {
+      array.splice(index, 1);
+    }
+    return array;
+  }
 }
